refactor: 신고 응답 DTO명 변경
