+++ src/main/java/coffeemeet/server/report/service/ReportService.java
@@ -8,8 +8,8 @@ import coffeemeet.server.common.execption.NotFoundException;
 import coffeemeet.server.report.domain.Report;
 import coffeemeet.server.report.implement.ReportCommand;
 import coffeemeet.server.report.implement.ReportQuery;
-import coffeemeet.server.report.service.dto.AllReportDto;
 import coffeemeet.server.report.service.dto.ReportDto;
+import coffeemeet.server.report.service.dto.ReportDetailDto;
 import coffeemeet.server.report.service.dto.TargetReportDto;
 import coffeemeet.server.user.domain.User;
 import coffeemeet.server.user.implement.UserQuery;
@@ -47,14 +47,14 @@ public class ReportService {
     reportCommand.save(report);
   }
 
-  public ReportDto.Response findReportById(Long reportId) {
+  public ReportDetailDto.Response findReportById(Long reportId) {
     Report report = reportQuery.getReportById(reportId);
     User reporter = userQuery.getUserById(report.getReporterId());
     User targetUser = userQuery.getUserById(report.getTargetId());
-    return ReportDto.Response.of(report, reporter, targetUser);
+    return ReportDetailDto.Response.of(report, reporter, targetUser);
   }
 
-  public List<AllReportDto.Response> findAllReports() {
+  public List<ReportDto.Response> findAllReports() {
     List<Report> allReports = reportQuery.getAllReports();
     return allReports.stream()
         .map(this::mapToAllReportDto)
@@ -88,9 +88,9 @@ public class ReportService {
     }
   }
 
-  private AllReportDto.Response mapToAllReportDto(Report report) {
+  private ReportDto.Response mapToAllReportDto(Report report) {
     User targetUser = userQuery.getUserById(report.getTargetId());
-    return AllReportDto.Response.of(targetUser);
+    return ReportDto.Response.of(targetUser, report);
   }
 
   private TargetReportDto.Response mapToReportDto(Report report) {
